feat: replace in memory CDC with task based state machine
