+++ core/transfer-process-manager/src/main/java/org/eclipse/edc/virtual/controlplane/transfer/process/TransferManagerExtension.java
@@ -24,6 +24,7 @@ import org.eclipse.edc.connector.controlplane.transfer.spi.store.TransferProcess
 import org.eclipse.edc.runtime.metamodel.annotation.Extension;
 import org.eclipse.edc.runtime.metamodel.annotation.Inject;
 import org.eclipse.edc.runtime.metamodel.annotation.Provider;
+import org.eclipse.edc.spi.command.CommandHandlerRegistry;
 import org.eclipse.edc.spi.message.RemoteMessageDispatcherRegistry;
 import org.eclipse.edc.spi.monitor.Monitor;
 import org.eclipse.edc.spi.security.Vault;
@@ -31,7 +32,14 @@ import org.eclipse.edc.spi.system.ServiceExtension;
 import org.eclipse.edc.spi.system.ServiceExtensionContext;
 import org.eclipse.edc.transaction.spi.TransactionContext;
 import org.eclipse.edc.virtual.controlplane.participantcontext.spi.ParticipantWebhookResolver;
+import org.eclipse.edc.virtual.controlplane.tasks.TaskService;
+import org.eclipse.edc.virtual.controlplane.transfer.process.task.TransferProcessTaskExecutorImpl;
+import org.eclipse.edc.virtual.controlplane.transfer.process.task.command.CompleteTransferCommandHandler;
+import org.eclipse.edc.virtual.controlplane.transfer.process.task.command.ResumeTransferCommandHandler;
+import org.eclipse.edc.virtual.controlplane.transfer.process.task.command.SuspendTransferCommandHandler;
+import org.eclipse.edc.virtual.controlplane.transfer.process.task.command.TerminateTransferCommandHandler;
 import org.eclipse.edc.virtual.controlplane.transfer.spi.TransferProcessStateMachineService;
+import org.eclipse.edc.virtual.controlplane.transfer.spi.TransferProcessTaskExecutor;
 
 import java.time.Clock;
 
@@ -76,7 +84,22 @@ public class TransferManagerExtension implements ServiceExtension {
     private Vault vault;
 
     @Inject
+    private TaskService taskService;
+
+    @Inject
     private Clock clock;
+
+    @Inject
+    private CommandHandlerRegistry commandHandlerRegistry;
+
+    // temporary workaround to register command handlers for overriding the default ones
+    @Override
+    public void prepare() {
+        commandHandlerRegistry.register(new SuspendTransferCommandHandler(store, taskService, clock));
+        commandHandlerRegistry.register(new ResumeTransferCommandHandler(store, taskService, clock));
+        commandHandlerRegistry.register(new TerminateTransferCommandHandler(store, taskService, clock));
+        commandHandlerRegistry.register(new CompleteTransferCommandHandler(store, taskService, clock));
+    }
 
     @Provider
     public TransferProcessManager transferProcessManager() {
@@ -97,6 +120,25 @@ public class TransferManagerExtension implements ServiceExtension {
                 .pendingGuard(pendingGuard)
                 .observable(observable)
                 .policyArchive(policyArchive)
+                .build();
+    }
+
+    @Provider
+    public TransferProcessTaskExecutor transferProcessTaskExecutor() {
+        return TransferProcessTaskExecutorImpl.Builder.newInstance()
+                .store(store)
+                .transactionContext(transactionContext)
+                .dataFlowController(dataFlowController)
+                .dispatcherRegistry(dispatcherRegistry)
+                .webhookResolver(webhookResolver)
+                .vault(vault)
+                .addressResolver(dataAddressResolver)
+                .monitor(monitor)
+                .pendingGuard(pendingGuard)
+                .observable(observable)
+                .policyArchive(policyArchive)
+                .taskService(taskService)
+                .clock(clock)
                 .build();
     }
