+++ core/transfer-process-manager/src/main/java/org/eclipse/edc/virtual/controlplane/transfer/process/task/TransferProcessTaskExecutorImpl.java
@@ -374,7 +374,10 @@ public class TransferProcessTaskExecutorImpl implements TransferProcessTaskExecu
     private void transitionToStarted(TransferProcess transferProcess) {
         transferProcess.transitionStarted();
         update(transferProcess);
-        observable.invokeForEach(l -> l.started(transferProcess, TransferProcessStartedData.Builder.newInstance().build()));
+        var transferStartedData = TransferProcessStartedData.Builder.newInstance()
+                .dataAddress(transferProcess.getContentDataAddress())
+                .build();
+        observable.invokeForEach(l -> l.started(transferProcess, transferStartedData));
     }
 
     private void transitionToCompleted(TransferProcess transferProcess) {
