--- core/transfer-process-task-executor/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
@@ -1,14 +1,0 @@
-#
-#  Copyright (c) 2026 Metaform Systems, Inc.
-#
-#  This program and the accompanying materials are made available under the
-#  terms of the Apache License, Version 2.0 which is available at
-#  https://www.apache.org/licenses/LICENSE-2.0
-#
-#  SPDX-License-Identifier: Apache-2.0
-#
-#  Contributors:
-#       Metaform Systems, Inc. - initial API and implementation
-#
-#
-org.eclipse.edc.virtual.controlplane.transfer.process.task.TransferTaskExecutorExtension
