--- extensions/cdc/memory/negotiation-cdc-memory/src/main/java/org/eclipse/edc/virtual/controlplane/contract/negotiation/store/ContractNegotiationChangeRegistry.java
@@ -1,24 +1,0 @@
-/*
- *  Copyright (c) 2025 Metaform Systems, Inc.
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Metaform Systems, Inc. - initial API and implementation
- *
- */
-
-package org.eclipse.edc.virtual.controlplane.contract.negotiation.store;
-
-import org.eclipse.edc.runtime.metamodel.annotation.ExtensionPoint;
-import org.eclipse.edc.virtual.controlplane.contract.spi.negotiation.ContractNegotiationChangeListener;
-
-@ExtensionPoint
-public interface ContractNegotiationChangeRegistry {
-
-    void registerChangeListener(ContractNegotiationChangeListener listener);
-}
