+++ settings.gradle.kts
@@ -35,9 +35,7 @@ include(":spi:iam:decentralized-claims:dcp-scope-spi")
 include(":core:v-connector-core")
 include(":core:v-task-core")
 include(":core:negotiation-manager")
-include(":core:negotiation-task-executor")
 include(":core:transfer-process-manager")
-include(":core:transfer-process-task-executor")
 
 // data-protocols
 include(":data-protocols:dsp")
@@ -54,8 +52,6 @@ include(":data-protocols:dsp:dsp-2025:dsp-transfer-process-2025:dsp-transfer-pro
 include(":extensions:cel:cel-spi")
 include(":extensions:cel:cel-extension")
 include(":extensions:cel:cel-store-sql")
-include(":extensions:cdc:memory:negotiation-cdc-memory")
-include(":extensions:cdc:memory:transfer-process-cdc-memory")
 include(":extensions:cdc:postgres:postgres-cdc")
 include(":extensions:cdc:publisher:negotiation-cdc-publisher-nats")
 include(":extensions:cdc:subscriber:negotiation-subscriber-nats")
@@ -101,7 +97,6 @@ include(":system-tests:dsp-tck-tests")
 include(":system-tests:extensions:v-tck-extension")
 include(":system-tests:extensions:v-tasks-tck-extension")
 include(":system-tests:runtimes:tck:tck-controlplane-memory")
-include(":system-tests:runtimes:tck:tck-controlplane-memory-tasks")
 include(":system-tests:runtimes:tck:tck-controlplane-postgres")
 include(":system-tests:runtimes:tck:tck-controlplane-postgres-tasks")
 include(":system-tests:runtimes:e2e:e2e-controlplane-memory")
@@ -118,5 +113,6 @@ include(":dist:bom:virtual-controlplane-cdc-base-bom")
 include(":dist:bom:virtual-controlplane-feature-dcp-bom")
 include(":dist:bom:virtual-controlplane-feature-sql-bom")
 include(":dist:bom:virtual-controlplane-feature-nats-bom")
+include(":dist:bom:virtual-controlplane-feature-nats-tasks-bom")
 include(":dist:bom:virtual-controlplane-feature-nats-cdc-bom")
