+++ system-tests/dsp-tck-tests/build.gradle.kts
@@ -31,7 +31,6 @@ dependencies {
     runtimeOnly(libs.parsson)
 
     testCompileOnly(project(":system-tests:runtimes:tck:tck-controlplane-memory"))
-    testCompileOnly(project(":system-tests:runtimes:tck:tck-controlplane-memory-tasks"))
     testCompileOnly(project(":system-tests:runtimes:tck:tck-controlplane-postgres"))
     testCompileOnly(project(":system-tests:runtimes:tck:tck-controlplane-postgres-tasks"))
