--- system-tests/dsp-tck-tests/src/test/java/org/eclipse/edc/virtual/tck/dsp/MemoryTasksEdcVirtualCompatibilityDockerTest.java
@@ -1,71 +1,0 @@
-/*
- *  Copyright (c) 2025 Metaform Systems, Inc.
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Metaform Systems, Inc. - initial API and implementation
- *
- */
-
-package org.eclipse.edc.virtual.tck.dsp;
-
-import org.eclipse.edc.junit.annotations.EndToEndTest;
-import org.eclipse.edc.junit.extensions.ComponentRuntimeExtension;
-import org.eclipse.edc.junit.extensions.RuntimeExtension;
-import org.eclipse.edc.spi.system.configuration.Config;
-import org.eclipse.edc.spi.system.configuration.ConfigFactory;
-import org.junit.jupiter.api.extension.RegisterExtension;
-import org.testcontainers.junit.jupiter.Testcontainers;
-
-import java.util.HashMap;
-import java.util.List;
-
-@EndToEndTest
-@Testcontainers
-public class MemoryTasksEdcVirtualCompatibilityDockerTest extends DspCompatibilityDockerTestBase {
-
-    protected static final String CONNECTOR = "CUT";
-
-
-    @RegisterExtension
-    protected static RuntimeExtension runtime = ComponentRuntimeExtension.Builder.newInstance()
-            .name(CONNECTOR)
-            .modules(":system-tests:runtimes:tck:tck-controlplane-memory-tasks")
-            .configurationProvider(MemoryTasksEdcVirtualCompatibilityDockerTest::runtimeConfiguration)
-            .build();
-
-
-    private static Config runtimeConfiguration() {
-        return ConfigFactory.fromMap(new HashMap<>() {
-            {
-                put("edc.participant.id", "participantContextId");
-                put("web.http.protocol.port", "8282"); // this must match the configured connector url in resources/docker.tck.properties
-                put("web.http.protocol.path", "/api/dsp"); // this must match the configured connector url in resources/docker.tck.properties
-                put("edc.dsp.callback.address", "http://host.docker.internal:8282/api/dsp"); // host.docker.internal is required by the container to communicate with the host
-                put("edc.hostname", "host.docker.internal");
-                put("edc.component.id", "DSP-compatibility-test");
-                put("edc.transfer.proxy.token.signer.privatekey.alias", "private-key");
-                put("edc.transfer.proxy.token.verifier.publickey.alias", "public-key");
-                put("edc.iam.oauth2.jwks.url", "https://example.com/jwks");
-                put("edc.iam.oauth2.issuer", "test-issuer");
-            }
-        });
-    }
-
-    @Override
-    protected List<String> getAllowedFailures() {
-        return List.of();
-    }
-
-    @Override
-    protected String dockerConfigFilePath() {
-        return resourceConfig("docker.tck.tasks.properties");
-    }
-
-}
-
