+++ system-tests/runtime-tests/src/test/java/org/eclipse/edc/virtual/RuntimeSmokeTests.java
@@ -165,6 +165,54 @@ public class RuntimeSmokeTests {
     @Nested
     @PostgresqlIntegrationTest
     @Testcontainers
+    class ControlPlaneTasksPgDcp extends SmokeTest {
+
+        @Order(0)
+        @RegisterExtension
+        static final NatsEndToEndExtension NATS_EXTENSION = new NatsEndToEndExtension();
+
+        @Order(0)
+        @RegisterExtension
+        static final PostgresqlEndToEndExtension POSTGRESQL_EXTENSION = new PostgresqlEndToEndExtension(createPgContainer());
+        static final String DB_NAME = "smoke_test";
+
+
+        @Order(1)
+        @RegisterExtension
+        static final BeforeAllCallback CREATE_DATABASES = context -> {
+            POSTGRESQL_EXTENSION.createDatabase(DB_NAME);
+        };
+        
+        @RegisterExtension
+        @Order(2)
+        static final RuntimeExtension RUNTIME = ComponentRuntimeExtension.Builder.newInstance()
+                .name("control-plane-pg")
+                .modules(":dist:bom:virtual-controlplane-memory-bom",
+                        ":dist:bom:virtual-controlplane-feature-dcp-bom",
+                        ":dist:bom:virtual-controlplane-feature-sql-bom",
+                        ":dist:bom:virtual-controlplane-feature-nats-tasks-bom")
+                .endpoints(ENDPOINTS.build())
+                .configurationProvider(RuntimeSmokeTests::config)
+                .configurationProvider(ControlPlaneTasksPgDcp::config)
+                .configurationProvider(NATS_EXTENSION::configFor)
+                .configurationProvider(() -> POSTGRESQL_EXTENSION.configFor(DB_NAME))
+                .paramProvider(DefaultEndpoint.class, RuntimeSmokeTests::defaultEndpoint)
+                .build();
+
+        private static Config config() {
+            return ConfigFactory.fromMap(new HashMap<>() {
+                {
+                    put("edc.iam.oauth2.issuer", "test-issuer");
+                    put("edc.iam.oauth2.jwks.cache.validity", "0");
+                    put("edc.iam.oauth2.jwks.url", "https://example.com/jwks");
+                }
+            });
+        }
+    }
+
+    @Nested
+    @PostgresqlIntegrationTest
+    @Testcontainers
     class CdcAgent extends SmokeTest {
 
         @Order(0)
