+++ system-tests/runtimes/e2e/e2e-controlplane-postgres-nats-tasks/build.gradle.kts
@@ -20,9 +20,6 @@ dependencies {
     implementation(libs.edc.iam.mock)
     implementation(project(":dist:bom:virtual-controlplane-base-bom"))
     implementation(project(":dist:bom:virtual-controlplane-feature-sql-bom"))
-    implementation(project(":core:transfer-process-task-executor"))
-    implementation(project(":core:negotiation-task-executor"))
-    implementation(project(":core:v-task-core"))
     implementation(project(":extensions:control-plane:tasks:listener:tasks-store-listener"))
     implementation(project(":extensions:control-plane:tasks:publisher:tasks-publisher-nats"))
     implementation(project(":extensions:control-plane:tasks:subscriber:negotiation-tasks-subscriber-nats"))
