+++ system-tests/runtimes/tck/tck-controlplane-memory/build.gradle.kts
@@ -17,10 +17,8 @@ plugins {
 }
 
 dependencies {
-    implementation(project(":system-tests:extensions:v-tck-extension"))
+    implementation(project(":system-tests:extensions:v-tasks-tck-extension"))
     implementation(project(":dist:bom:virtual-controlplane-memory-bom"))
-    runtimeOnly(libs.edc.tck.extension)
-    runtimeOnly(libs.edc.core.participantcontext.single)
     runtimeOnly(libs.edc.bom.dataplane) {
         exclude(module = "data-plane-selector-client")
     }
