+++ system-tests/runtimes/tck/tck-controlplane-postgres-tasks/build.gradle.kts
@@ -21,9 +21,6 @@ dependencies {
     implementation(project(":system-tests:extensions:v-tasks-tck-extension"))
     implementation(project(":dist:bom:virtual-controlplane-base-bom"))
     implementation(project(":dist:bom:virtual-controlplane-feature-sql-bom"))
-    implementation(project(":core:transfer-process-task-executor"))
-    implementation(project(":core:negotiation-task-executor"))
-    implementation(project(":core:v-task-core"))
     implementation(project(":extensions:control-plane:tasks:listener:tasks-store-listener"))
     implementation(project(":extensions:control-plane:tasks:publisher:tasks-publisher-nats"))
     implementation(project(":extensions:control-plane:tasks:subscriber:negotiation-tasks-subscriber-nats"))
